backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
pymysql
boto3
openai
python-jose[cryptography]

requests
python-dotenv
langchain-community
langchain-openai
langchain
langgraph
openpyxl
pandas
numpy